\name{AFFINEpack-package}
\alias{AFFINEpack-package}
\alias{AFFINEpack}
\docType{package}
\title{
\packageTitle{AFFINEpack}
}
\description{
\packageDescription{AFFINEpack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AFFINEpack}
\packageIndices{AFFINEpack}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{AFFINEpack}

Maintainer: \packageMaintainer{AFFINEpack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
